+++ Libraries/aaronbrethorst-onebusaway-iphone-common-fa9a21b/Classes/View/OBARequestDrivenTableViewController.m
@@ -165,7 +165,7 @@
     cell.textLabel.text = @"Updating...";
     cell.textLabel.font = [UIFont systemFontOfSize:18];
     cell.textLabel.textColor = [UIColor grayColor];
-    cell.textLabel.textAlignment = UITextAlignmentCenter;    
+    cell.textLabel.textAlignment = NSTextAlignmentCenter;    
     
     return cell;
 }